refactor: fix duplicate child being spawned, process not detecting socket reloads and remove babel
