--- .babelrc
@@ -1,15 +1,0 @@
-{
-  "presets": [
-    ["env", {
-      "targets": {
-        "node": "current"
-      }
-    }]
-  ],
-  "plugins": [
-    ["module-resolver", {
-      "root": ["./src"]
-    }],
-    ["transform-object-rest-spread", "transform-class-properties", "syntax-dynamic-import", "syntax-object-rest-spread"]
-  ]
-}
