+++ package.json
@@ -3,11 +3,13 @@
   "version": "1.0.0",
   "description": "Enables the consumption of docker events as a node event emitter",
   "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "files": [
     "dist"
   ],
   "scripts": {
-    "build": "rm -rf dist/* && babel src -d dist",
+    "prepare": "npm run build",
+    "build": "del-cli dist && tsup ./src/index.ts --dts",
     "lint": "eslint src __tests__",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -27,22 +29,19 @@
   },
   "homepage": "https://github.com/GridPlus/docker-events#readme",
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.3",
-    "babel-eslint": "^8.2.6",
-    "babel-plugin-module-resolver": "^3.1.1",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-preset-env": "^1.7.0",
+    "@types/debug": "^4.1.6",
+    "@types/node": "^16.3.1",
+    "del-cli": "^4.0.0",
     "eslint": "^5.3.0",
     "eslint-config-airbnb": "^17.1.0",
     "eslint-plugin-import": "^2.14.0",
     "eslint-plugin-jest": "^21.21.0",
     "eslint-plugin-jsx-a11y": "^6.1.1",
-    "eslint-plugin-react": "^7.11.0"
+    "eslint-plugin-react": "^7.11.0",
+    "tsup": "^4.12.5",
+    "typescript": "^4.3.5"
   },
   "dependencies": {
-    "debug": "^3.1.0",
-    "llog": "^0.1.1",
-    "pino": "^5.0.4"
+    "debug": "^3.1.0"
   }
 }
